ai_agent_service/requirements.txt
azure-ai-projects
azure-ai-agents==1.1.0b4
azure-identity